requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==4.0.1
boto3==1.42.54
botocore==1.42.54
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cmdstanpy==1.3.0
contourpy==1.3.3
coverage==7.13.4
cryptography==42.0.8
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
environs==14.6.0
et_xmlfile==2.0.0
fastapi==0.115.0
fastapi_cors==0.0.1
fonttools==4.61.1
frozenlist==1.8.0
h11==0.16.0
holidays==0.86
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
importlib_resources==6.5.2
iniconfig==2.3.0
jmespath==1.1.0
joblib==1.5.3
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.2.2
matplotlib==3.10.8
multidict==6.7.1
numpy==1.26.4
oauthlib==3.3.1
openpyxl==3.1.2
packaging==26.0
pandas==2.2.2
passlib==1.7.4
pillow==12.1.1
pluggy==1.6.0
propcache==0.4.1
prophet==1.1.5
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.2.1
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
pytest==8.2.1
pytest-cov==5.0.0
python-dateutil==2.9.0
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.3
redis==7.2.0
reportlab==4.4.10
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.16.0
scikit-learn==1.5.0
scipy==1.17.0
sentry-sdk==2.5.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stanio==0.5.1
starlette==0.38.6
stripe==9.0.0
threadpoolctl==3.6.0
tqdm==4.67.3
twilio==9.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.30.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yarl==1.22.0
